rename pyslavseq to src
