+++ workflow/scripts/folds.py
@@ -10,10 +10,10 @@ import functools
 import re
 import numpy as np
 from pathlib import Path
-from pyslavseq.genome.Genome import Genome
-from pyslavseq.genome.Interval import Interval, is_within
-import pyslavseq.genome.interval_generator as ig
-from get_windows import read_rmsk, make_l1_windows
+from src.genome.Genome import Genome
+from src.genome.Interval import Interval, is_within
+from src.genome import interval_generator as ig
+from src.genome.windows import read_rmsk, make_l1_windows
 
 def interval_hash(iv, window_size):
     return str(iv.chrom) + "_" + str(iv.start // window_size)
