+++ workflow/scripts/get_eul1db.py
@@ -4,7 +4,7 @@ __author__ = 'Michael Cuoco'
 
 import pandas as pd
 import sys,gc,traceback
-from pyslavseq.genome import Interval, Genome
+from src.genome import Interval, Genome
     
 def main():
     '''
