+++ workflow/scripts/src/features/occupied_windows.py
